// frontend/src/api/generated/models/sizingConfigResponse.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * vibe-quant API
 * OpenAPI spec version: 0.1.0
 */
import type { SizingConfigResponseConfig } from './sizingConfigResponseConfig';

export interface SizingConfigResponse {
  id: number;
  name: string;
  method: string;
  config: SizingConfigResponseConfig;
  created_at: string;
}